Django/requirements-local.txt
aniso8601==6.0.0
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.1.1
django-cors-headers==3.5.0
django-graphql-jwt==0.3.1
environ==1.0
filelock==3.0.12
graphene==2.1.7
graphene-django==2.13.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==20.0.4
idna==2.10
mysqlclient==1.4.6
oauthlib==3.1.0
promise==2.2.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2019.2
requests==2.24.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.11
virtualenv==20.0.20
gunicorn==20.0.4
